// src/enums/route.enum.ts
export enum MetadataRoutes {
  HOME = 'home',
  LOGIN = 'login',
  PRICING = 'pricing',
  CONTACT_US = 'contact',
  PRODUCTS = 'products',
  UPI_QR_COLLECTIONS = 'upi-qr-collections',
  MONEY_TRANSFER = 'money-transfer',
  POS_DEVICES = 'pos-devices',
  DYNAMIC_QR_COLLECTIONS = 'dynamic-qr-collections',
  INSTANT_BANKING = 'instant-banking',
  PRIVACY_POLICY = 'privacy-policy',
  TERMS_AND_CONDITIONS = 'terms-and-conditions',
}